fix tests
minor fixes in vault factory tests
bypass 1 test suite of unused file
